+++ src/Domain/Models/PlanModel/Plan.cs
@@ -9,7 +9,7 @@ namespace Domain.Models.PlanModel
     {
         public Guid Id { get; set; } = Guid.NewGuid();
         public string TypePlan { get; set; }
-        public float ValuePlan { get; set; }
+        public double ValuePlan { get; set; }
         public string TransactionWay { get; set; }
         public string UserID { get; set; }
         public bool IsActive { get; set; }
