+++ src/Infrastructure/Migrations/ApplicationDataContextModelSnapshot.cs
@@ -73,7 +73,7 @@ namespace Infrastructure.Migrations
                     b.Property<DateTime>("ReturnDeadLine")
                         .HasColumnType("TEXT");
 
-                    b.Property<float>("ReturnRate")
+                    b.Property<double>("ReturnRate")
                         .HasColumnType("REAL");
 
                     b.Property<string>("TransactionWay")
@@ -88,7 +88,7 @@ namespace Infrastructure.Migrations
                         .IsRequired()
                         .HasColumnType("TEXT");
 
-                    b.Property<float>("ValuePlan")
+                    b.Property<double>("ValuePlan")
                         .HasColumnType("REAL");
 
                     b.HasKey("Id");
@@ -102,7 +102,7 @@ namespace Infrastructure.Migrations
                         .ValueGeneratedOnAdd()
                         .HasColumnType("TEXT");
 
-                    b.Property<float>("InterestRate")
+                    b.Property<double>("InterestRate")
                         .HasColumnType("REAL");
 
                     b.Property<bool>("IsActive")
@@ -123,7 +123,7 @@ namespace Infrastructure.Migrations
                         .IsRequired()
                         .HasColumnType("TEXT");
 
-                    b.Property<float>("ValuePlan")
+                    b.Property<double>("ValuePlan")
                         .HasColumnType("REAL");
 
                     b.HasKey("Id");
